Search is a plugin: the matcher and the header box become a row (13)
`@olai/index` and `Query.search` become `packages/plugins/search/`. Core
defines a `Search` door in `@olai/ops`; `search.nodes` and the
`search_nodes` tool call through it and refuse in words when nobody is
mounted. The header's box is the row's browser half, in a new `lead` seat
of `app.header`; the shortlist kit four core doors share stays core.
