+++ docs/index.md
@@ -24,6 +24,7 @@ The first four are **engines** — the ACP agents the chat panel can seat, one p
 - [plugins/odu.md](plugins/odu.md) — CI on the board: the chip a `worktree` wears while odu is running in that checkout, the run matrix it opens, what a settled run leaves behind, the two facts on the board that turn it on, the doorbell that wakes a scoped conversation when a claimed run goes red or settles, the chat panel's `odu mcp`, and what the watching costs.
 - [plugins/xyne-spaces.md](plugins/xyne-spaces.md) — the fleet in team chat: doorbell digests and trimmed orchestrator replies mirrored into a bound Xyne Spaces channel, a live working signal while a turn runs, watch-only (nothing comes back), env-held secrets and a `xyne-channel` property on the node agent.
 - [plugins/git.md](plugins/git.md) — the ledger as a row: what `--plugins` without it leaves (writes land, recorded by nobody, no pill), the pin (`--commit` / `--push`), and the sibling cells an agent reads.
+- [plugins/search.md](plugins/search.md) — the matcher as a row: what `--plugins` without it leaves (the grammar, the tool and every box still there, and every query answered with the reason), why it composes nothing on the wire, the bar's `lead` seat, and why the filter over a page is not on this row.
 - [dynamic-plugins.md](dynamic-plugins.md) — the other kind of plugin: one somebody writes INTO the served directory, as a node with a `plugin` property and its two halves in child notes, which olai compiles and mounts while it runs. What a half may import and what it may not, why the approval is a property on the plugin's own node and names a version, the three tools an agent gets and the two verbs it does not, and why the code runs with the server's authority.
 
 The development docs — the roadmap in the format itself, the decisions and why the alternatives lost, and the root-cause analyses — live in the orchestrator's own vault, [juspay/oss.olai](https://github.com/juspay/oss.olai), under its `olai/` folder:
