+++ docs/plugins/search.md
@@ -1,0 +1,1 @@
+../../packages/plugins/search/docs.md
