+++ packages/bundle/package.json
@@ -17,7 +17,7 @@
     "typecheck": "tsc --noEmit"
   },
   "//": "WHAT A PLUGIN'S NAME MAY BE SPELLED IN — olai.yml is the build's rows and the only hand-written spelling; generate.ts writes the other three (the browser's rows, the stylesheet chain and the testid merge) out of it; and src/fence.test.ts is what refuses a general package that names a tenant before a reviewer has to notice.",
-  "//dependencies": "Every plugin this build has — four ENGINES and three TENANTS — the interface they are written against, and the bridge that mounts them. The plugin dependencies are declared even though nothing here imports one STATICALLY any more: a dynamic `import()` and a loader-resolved specifier are both real edges, and a manifest that did not declare them would be a package whose install could not resolve its own generated rows. `js-yaml` is NOT declared and is not an omission — it is the cordis pin.s, read from the root the way every hydrated package.s runtime dependency is, and `../generate.ts` reads the rows with the same one. `@olai/plugin-api` is the INTERFACE and is a dependency in this direction only — it names no plugin, which is what makes the arrow one-way and the cycle unrepresentable. `@olai/effect-cordis` is the plugin RUNTIME, and this package is one of the two that name it: mounting a bundle is a loader entry and reading a row's state is a registry walk, and both of those are Cordis behind a door that says Effect. `cordis` and `@cordisjs/plugin-loader`/`-include` are NOT named here at all — src/fence.test.ts refuses them in every package but the bridge — and they are hydrated from the npins pin (nix/cordis.nix) into the root node_modules, the way every `@kolu/*` member is, with scripts/check-hydrated-deps.sh holding their versions. `effect` is for the schemas a contributed member is declared in.",
+  "//dependencies": "Every plugin this build has — four ENGINES, three TENANTS and the four core rows — the interface they are written against, and the bridge that mounts them. The plugin dependencies are declared even though nothing here imports one STATICALLY any more: a dynamic `import()` and a loader-resolved specifier are both real edges, and a manifest that did not declare them would be a package whose install could not resolve its own generated rows. `js-yaml` is NOT declared and is not an omission — it is the cordis pin.s, read from the root the way every hydrated package.s runtime dependency is, and `../generate.ts` reads the rows with the same one. `@olai/plugin-api` is the INTERFACE and is a dependency in this direction only — it names no plugin, which is what makes the arrow one-way and the cycle unrepresentable. `@olai/effect-cordis` is the plugin RUNTIME, and this package is one of the two that name it: mounting a bundle is a loader entry and reading a row's state is a registry walk, and both of those are Cordis behind a door that says Effect. `cordis` and `@cordisjs/plugin-loader`/`-include` are NOT named here at all — src/fence.test.ts refuses them in every package but the bridge — and they are hydrated from the npins pin (nix/cordis.nix) into the root node_modules, the way every `@kolu/*` member is, with scripts/check-hydrated-deps.sh holding their versions. `effect` is for the schemas a contributed member is declared in.",
   "dependencies": {
     "@olai/effect-cordis": "workspace:*",
     "@olai/plugin-api": "workspace:*",
@@ -30,6 +30,7 @@
     "olai-plugin-odu": "workspace:*",
     "olai-plugin-opencode": "workspace:*",
     "olai-plugin-pi": "workspace:*",
+    "olai-plugin-search": "workspace:*",
     "olai-plugin-xyne-spaces": "workspace:*",
     "olai-plugin-git": "workspace:*"
   }
