--- packages/index/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@olai/index",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "scripts": {
-    "typecheck": "tsc --noEmit"
-  },
-  "//dependencies": "@olai/format and nothing else. This package holds the index and no rules: what a query MEANS, which fields a word is looked for in and what order an answer comes back in are all one layer down, and the only things reached for are the three that make a superset sayable — `hayOf` / `documentHayOf` (the text to hold) and `narrowableBy` (which of a query's words may narrow it). The engine is `bun:sqlite`, which is the runtime's own and therefore not a line here; that it is Bun's is why this sits above `format` rather than inside it, since `format` is bundled into a browser.",
-  "dependencies": {
-    "@olai/format": "workspace:*"
-  },
-  "//devDependencies": "@types/bun is the root's; nothing else is needed. The tests and the bench build their corpora with @olai/format's own `./testlib`, which the runtime dependency above already resolves.",
-  "devDependencies": {}
-}
