--- packages/index/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  // Per-package `tsc --noEmit` anchor. Shared options live in
-  // ../../tsconfig.base.json; each member only points tsc at its own src.
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "rootDir": "src"
-  },
-  "include": ["src"]
-}
