+++ packages/ops/src/tools.test.ts
@@ -49,134 +49,23 @@ import {
   CreateRequest,
   NodeAnswer,
   NodeRequest,
-  type OutlineSet,
   type Placed,
   type Placement,
-  type Reading,
-  type SearchRequest,
   type Subtree,
 } from "@olai/format"
 import { expect, test } from "bun:test"
 import { Effect, Schema } from "effect"
-import { NO_KINDS } from "@olai/format"
 
-import { readingOf, setOf, steady } from "./fixtures.testlib.ts"
-import { act, asking, read, TOOLS, write } from "./tools.ts"
+import { NO_SEARCH } from "./ops.ts"
+import { act, read, TOOLS, write } from "./tools.ts"
+import { answerOf, CALLS, gaveOf, READS } from "./tools.testlib.ts"
 
-/** One house, and everything a read can carry: both marker kinds, a note, a
- *  date, both tag sigils, a placement with a parent and one without, a child
- *  deep enough to truncate a walk, and a file that does not parse — plus the
- *  documents beside it, which the two document reads answer over: one with a
- *  heading, one with none, a `.html` the set keeps no body for, and a `.md`
- *  that could not be read. */
-const EVERYTHING = (): OutlineSet =>
-  setOf({
-    "house.olai": [
-      `{"id":"house","ord":"a0","title":"House #home @sam","desc":"the note","date":"2026-08-14","doing":true,"see":["paint"],"after":["paint"]}`,
-      `{"id":"paint","parent":"house","ord":"a0","title":"paint the hall","done":"2026-08-09T10:15:00-04:00"}`,
-      `{"id":"sand","parent":"house","ord":"a1","title":"sand the floor","todo":true}`,
-      `{"id":"grain","parent":"sand","ord":"a0","title":"with the grain"}`,
-      // Under a node, so `house` has a `placed` row…
-      `{"id":"in-house","parent":"house","ord":"a2","mirror":"paint"}`,
-      // …and at the top level, so one of `paint`'s placements has no parent.
-      `{"id":"loose","ord":"a1","mirror":"paint"}`,
-    ].join("\n"),
-  }, [
-    ["notes/finishes.md", "# Finishes\n\nDoors: matte.\n"],
-    [
-      "notes/plan.md",
-      "---\nagent: claude-opus\nowners: [alice, bob]\n---\n# The plan\n",
-    ],
-    ["plain.md", "walnut, or birch\n"],
-    // Bare, because the set holds this one's PATH and not its content — which
-    // is exactly why no document read answers it.
-    "saved/page.html",
-  ], { "torn.olai": "{ not a record", "torn.md": "{ not a record" })
-
-const at = (): Reading => readingOf(EVERYTHING())
-
-/**
- * The read door, over that fixture — the SAME `asking` the ops layer builds
- * over its own gated read, so what this walks is the envelope an agent
- * actually receives and not a `Query` call the envelope is made of.
- *
- * `Effect.sync` rather than `succeed` so each question gets its own set,
- * exactly as the per-call `at()` this replaced did. Nothing here can fail —
- * the read is a fixture — so every answer is `runSync`-able.
- */
-const ASKING = asking(Effect.sync(at), steady().now, NO_KINDS)
-
-/** One read, answered. The tools' own effects never fail over a fixture that
- *  loaded, so the failure channel is discharged here rather than threaded
- *  through three tests that have nothing to say about it. */
-const answerOf = (
-  tool: Extract<(typeof TOOLS)[number], { kind: "read" }>,
-  args: unknown,
-): unknown => Effect.runSync(Effect.orDie(tool.ask(ASKING, args as never)))
-
-/**
- * What each read is CALLED with, one entry per tool and several calls per
- * entry.
- *
- * Several, because a read's answer is not one shape: `read_node` answers a
- * detail or the id it does not hold, a search answers hits or a refusal, and a
- * walk answers truncated or finished. Each of those is a decode of its own.
- */
-const CALLS: Record<string, ReadonlyArray<unknown>> = {
-  list_outlines: [{}],
-  search_nodes: [
-    { text: "hall" },
-    { text: "is:open" },
-    { text: "" },
-    { text: "date:today" },
-    { text: `"paint the hall" OR nothing-is-called-this` },
-    // The one field of a record a hit does not carry unless it is asked for.
-    { text: "House", withDesc: true },
-  ],
-  read_node: [
-    { id: "house" },
-    { id: "paint" },
-    { id: "shed" },
-    // The caller SHAPED its children — the one new arm of the answer.
-    { id: "house", fields: ["title", "status", "done"] },
-  ],
-  // All three arms of the answer: one node walked, the WHOLE outline walked,
-  // and an id the set does not hold — and both walks SHAPED, since `fields`
-  // joined the answer's union.
-  read_subtree: [
-    { id: "house", depth: 1 },
-    { id: "house" },
-    { file: "house.olai" },
-    { file: "house.olai", depth: 1 },
-    { id: "shed" },
-    // The lean walk: notes off, structure on. Reached here so the decode
-    // above is not vacuous about a `withDesc: false` answer.
-    { id: "house", withDesc: false },
-    // The shaped walks: one per way in, so both projected arms of the union
-    // are actually decoded.
-    { id: "house", fields: ["title", "status", "done"] },
-    { file: "house.olai", depth: 1, fields: ["title", "status", "custom"] },
-  ],
-  list_documents: [{}],
-  // The reads that REFUSE are not called here: this walk decodes ANSWERS, and
-  // a refusal has none. What `read_document` says about a path the set does not
-  // hold — and what `read_subtree` says about one, and about a call naming both
-  // ways in or neither — is the MCP face's own test and
-  // `an_external_agent.feature`'s, where the refusal travels as a tool result
-  // rather than being discharged by an `orDie` that would simply throw.
-  read_document: [{ file: "notes/finishes.md" }, { file: "plain.md" }],
-}
-
-const READS = TOOLS.filter((tool) => tool.kind === "read")
-
-/** Every answer the fixture can provoke, paired with the tool that gave it. */
-const answered = (): ReadonlyArray<{ name: string; answer: unknown }> =>
-  READS.flatMap((tool) =>
-    (CALLS[tool.name] ?? []).map((args) => ({
-      name: tool.name,
-      answer: tool.kind === "read" ? answerOf(tool, args) : undefined,
-    }))
-  )
+/** WITH NO MATCHER MOUNTED, which is what this package can build: the walk
+ *  is a harness now ({@link ./tools.testlib.ts}) and `olai-plugin-search`'s own
+ *  `tools.test.ts` runs the identical one with the real matcher behind
+ *  `search_nodes`. What is asserted here is every read core answers for
+ *  itself, plus the refusal envelope a serve without that row hands an agent. */
+const answers = () => gaveOf(NO_SEARCH)
 
 test("every read in the table is called here", () => {
   // The closure, and the reason the fixtures are a lookup rather than a list:
@@ -193,7 +82,7 @@ test("every answer decodes through the shape its own entry declares", () => {
       { errors: "all", onExcessProperty: "error" },
     )
     for (const args of CALLS[tool.name] ?? []) {
-      const answer = answerOf(tool, args)
+      const answer = answerOf(NO_SEARCH, tool, args)
       // Compared with what went in, so the assertion is "this IS the shape"
       // rather than "this parses" — a decode that dropped a field would
       // otherwise pass.
@@ -203,11 +92,7 @@ test("every answer decodes through the shape its own entry declares", () => {
 })
 
 test("the fixture reaches every optional field, so the check is not vacuous", () => {
-  const answers = answered()
-  const of = (name: string): ReadonlyArray<Record<string, unknown>> =>
-    answers.filter((one) => one.name === name).map((one) =>
-      one.answer as Record<string, unknown>
-    )
+  const of = answers()
 
   const outlines = of("list_outlines")[0]?.["outlines"] as ReadonlyArray<
     Record<string, unknown>
@@ -228,27 +113,21 @@ test("the fixture reaches every optional field, so the check is not vacuous", ()
     unreadable: [expect.any(String)],
   })
 
-  const searches = of("search_nodes")
-  expect(searches[0]?.["hits"]).toMatchObject([
-    { id: "paint", matched: "title", parent: "house" },
-  ])
-  // A query the grammar could not read carries the reason rather than an
-  // empty list with nothing to say.
-  expect(searches[1]?.["refusals"]).toBeArrayOfSize(1)
-  // And a relative word is counted from THIS LAYER'S CLOCK — the one a `done`
-  // is stamped with, which is what the door hands the grammar. `paint` was
-  // finished at the fixture clock's own instant, so `date:today` finds it: a
-  // door that read a clock of its own would answer this with nothing.
-  expect(searches[3]?.["hits"]).toMatchObject([{ id: "paint" }])
-  expect(searches[3]).not.toHaveProperty("refusals")
-  // A quoted PHRASE and an `OR` group reach this door too — the tool's own
-  // `text` prose spells both, and a schema that decoded the words and not
-  // these would be the door advertising a grammar it does not answer.
-  expect(searches[4]?.["hits"]).toMatchObject([{ id: "paint", matched: "title" }])
-  // And the note, when the query asked for it — the one field of a record a hit
-  // does not carry by default, reached here so the decode above is not vacuous
-  // about it.
-  expect(searches[5]?.["hits"]).toMatchObject([{ id: "house", desc: "the note" }])
+  // WHAT AN AGENT IS TOLD WHEN NOBODY IS SEARCHING — the same six queries,
+  // through the same tool, over a serve that mounts no matcher. Every one of
+  // them is no hits AND THE REASON, quoting what the reader typed: the shape a
+  // door already draws for a query the grammar could not read, spent on a
+  // serve that has no grammar to read it with.
+  //
+  // What each of those six queries FINDS is `olai-plugin-search`'s
+  // `tools.test.ts`, which runs this identical walk with the real matcher
+  // behind the door — the maximality of the hit shapes travelled there with
+  // the function that produces them.
+  for (const answer of of("search_nodes")) {
+    expect(answer["hits"]).toEqual([])
+    expect(answer["total"]).toBe(0)
+    expect(answer["refusals"]).toBeArrayOfSize(1)
+  }
 
   const [house, paint, gone] = of("read_node")
   expect(house).toMatchObject({
