+++ packages/plugins/search/README.md
