+++ packages/plugins/search/src/browser/all.css
@@ -1,0 +1,2 @@
+/* THIS PLUGIN'S SHEET — Tailwind emits only the utilities it can SEE. */
+@source "./";
