+++ packages/plugins/search/src/index.ts
@@ -1,0 +1,21 @@
+/**
+ * THE PLUGIN'S WORD, and the whole of this door.
+ *
+ * A tenant's root entry is its WIRE IDENTITY — the sibling surface it composes
+ * and the faces it exposes on it. This row composes none, and that is a ruling
+ * rather than an omission: `search.nodes` is CORE's member, on the browser, MCP
+ * and agent faces, and a row that published `surface/search/` of its own would
+ * rename the tool an agent already calls (`search_nodes`) for nothing. What this
+ * row contributes is the ANSWER behind that member — core defines the `Search`
+ * door and this plugin stands behind it — so nothing about it is on the wire.
+ *
+ * So what is left is the name, and it is here rather than in either half
+ * because BOTH halves need it and neither may be the other's source:
+ * `./server.ts` is mounted by the row and `./browser.tsx` is the chunk the tab
+ * fetches, and they share no graph at all. It IS the row's `id` — the word the
+ * fiber is bound under, the word `--plugins` takes, and the address of this
+ * plugin's docs page — and `@olai/bundle`'s `composition.test.ts` holds the two
+ * equal by loading the module its row names and reading this.
+ */
+
+export const name = "search"
